requirements.txt
python-json-logger==0.1.11
PyYAML==5.3.1
requests==2.24.0
redis==3.5.2
coverage==5.2.1
ptvsd==4.3.2